OAuth logic is moved to a separate controller package
