+++ src/main/java/com/example/authenticationserver/command/api/controller/oauth/OAuthLoginViewController.java
@@ -1,40 +1,32 @@
-package com.example.authenticationserver.command.api.controller;
+package com.example.authenticationserver.command.api.controller.oauth;
 
 import com.example.authenticationserver.security.AuthUserProfileProviderImpl;
 import com.example.authenticationserver.security.UserProfileDetails;
-import com.example.authenticationserver.service.UserProfileDetailsService;
 import com.project.core.commands.code.GenerateAuthorizationCodeCommand;
 import jakarta.servlet.http.HttpSession;
-import lombok.extern.slf4j.Slf4j;
 import org.axonframework.commandhandling.gateway.CommandGateway;
 import org.axonframework.queryhandling.QueryGateway;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.util.concurrent.CompletableFuture;
 
-@Slf4j
 @Controller
-@RequestMapping("/login")
-public class UserLoginViewController {
-  //TODO Add the ability to log in without additional parameters
-  private final QueryGateway queryGateway;
-  private final PasswordEncoder passwordEncoder;
+@RequestMapping("/login/oauth2/authorization")
+public class OAuthLoginViewController {
   private final CommandGateway commandGateway;
-  private final UserProfileDetailsService userProfileDetailsService;
   private final AuthUserProfileProviderImpl authUserProfileProvider;
   @Autowired
-  public UserLoginViewController(QueryGateway queryGateway, PasswordEncoder passwordEncoder, CommandGateway commandGateway, UserProfileDetailsService userProfileDetailsService, AuthUserProfileProviderImpl authUserProfileProvider) {
-    this.queryGateway = queryGateway;
-    this.passwordEncoder = passwordEncoder;
+  public OAuthLoginViewController(CommandGateway commandGateway, AuthUserProfileProviderImpl authUserProfileProvider) {
     this.commandGateway = commandGateway;
-    this.userProfileDetailsService = userProfileDetailsService;
     this.authUserProfileProvider = authUserProfileProvider;
   }
 
@@ -82,7 +74,7 @@ public class UserLoginViewController {
 
       return "redirect:" + builder.toUriString();
     } catch (Exception e) {
-      UriComponentsBuilder builder = UriComponentsBuilder.fromPath("/login")
+      UriComponentsBuilder builder = UriComponentsBuilder.fromPath("/login/oauth2/authorization")
         .queryParam("client_id", clientId)
         .queryParam("response_type", responseType)
         .queryParam("state", state)
@@ -93,23 +85,4 @@ public class UserLoginViewController {
       return "redirect:" + builder.toUriString();
     }
   }
-//  @PostMapping
-//  public String loginWithoutParams(@RequestParam("username") String username,
-//                                   @RequestParam("password") String password) {
-//    try {
-//      var auth = authUserProfileProvider.authenticate(new UsernamePasswordAuthenticationToken(username, password));
-//
-//      UserProfileDetails userProfileDetails = (UserProfileDetails) auth.getPrincipal();
-//
-//      UriComponentsBuilder builder = UriComponentsBuilder.fromPath("/security/user")
-//        .queryParam("error", "Invalid username or password");
-//
-//      return "redirect:" + builder.toUriString();
-//    } catch (Exception e) {
-//      UriComponentsBuilder builder = UriComponentsBuilder.fromPath("/login")
-//        .queryParam("error", "Invalid username or password");
-//
-//      return "redirect:" + builder.toUriString();
-//    }
-//  }
 }
