+++ src/main/java/com/example/authenticationserver/command/api/controller/oauth/OAuthRegistrationViewController.java
@@ -1,11 +1,8 @@
-package com.example.authenticationserver.command.api.controller;
+package com.example.authenticationserver.command.api.controller.oauth;
 
 import com.example.authenticationserver.command.api.service.UserProfileCommandService;
-import com.project.core.commands.user.CreateUserProfileCommand;
-import org.axonframework.commandhandling.gateway.CommandGateway;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,14 +10,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.util.UUID;
 @Controller
-@RequestMapping("/registration")
-public class UserRegistrationViewController {
+@RequestMapping("/registration/oauth2/authorization")
+public class OAuthRegistrationViewController {
   private final UserProfileCommandService userProfileCommandService;
   @Autowired
-  public UserRegistrationViewController(UserProfileCommandService userProfileCommandService) {
+  public OAuthRegistrationViewController(UserProfileCommandService userProfileCommandService) {
     this.userProfileCommandService = userProfileCommandService;
   }
 
@@ -50,7 +45,7 @@ public class UserRegistrationViewController {
 
     userProfileCommandService.register(new UsernamePasswordAuthenticationToken(username, password)).join();
 
-    UriComponentsBuilder builder = UriComponentsBuilder.fromPath("/login")
+    UriComponentsBuilder builder = UriComponentsBuilder.fromPath("/login/oauth2/authorization")
       .queryParam("client_id", clientId)
       .queryParam("response_type", responseType)
       .queryParam("state", state)
