+++ src/main/resources/templates/registration.html
@@ -8,7 +8,7 @@
 <body>
 <h2>Регистрация</h2>
 
-<form action="/registration" method="post">
+<form action="/registration/oauth2/authorization" method="post">
   <input type="hidden" name="client_id" th:value="${client_id}">
   <input type="hidden" name="response_type" th:value="${response_type}">
   <input type="hidden" name="state" th:value="${state}">
@@ -24,6 +24,6 @@
   </div>
   <button type="submit">Зарегистрироваться</button>
 </form>
-<p>Уже есть аккаунт? <a th:href="@{/login(client_id=${client_id}, response_type=${response_type}, state=${state}, scope=${scope}, redirect_url=${redirect_url})}">Войти</a></p>
+<p>Уже есть аккаунт? <a th:href="@{/login/oauth2/authorization(client_id=${client_id}, response_type=${response_type}, state=${state}, scope=${scope}, redirect_url=${redirect_url})}">Войти</a></p>
 </body>
 </html>
